The data type of the seed variety has been changed

+++ ShapeEngine/StaticLib/PerlinNoise2D.cs
@@ -10,8 +10,8 @@ namespace ShapeEngine.StaticLib
     public static class PerlinNoise2D
     {
         private static Dictionary<(int, int), Vector2> gradients = new Dictionary<(int, int), Vector2>();
-        private static int _seed = 0;
-        public static int seed
+        private static long _seed = 0;
+        public static long seed
         {
             get { return _seed; }
             set
